refactor: validator.js로 유효성 검증 로직 함수 분리
